% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbox.R
\name{gbox}
\alias{gbox}
\title{Create a quick box plot in ggplot.}
\usage{
gbox(x, y, hoz)
}
\arguments{
\item{x}{This is the vector to be plotted.}

\item{y}{This is the categorical vector (if there is one) to be plotted as well.}

\item{orientation}{This is a parameter that will dictate whether the plot is vertical or horizontal.}
}
\value{
This function returns a ggplot box plot object.
}
\description{
This will graph two given vectors in a ggplot-style box plot with the
x-axis labeled "x" and the y-axis labeled "y".
}
\examples{
## Create a box plot of y vs x.
x <- (1:10)
y <- ('red','red','red','blue','blue','blue','green','green','green','orange')
gbox(x, y, hoz)

}
