% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph.R
\name{ggraph}
\alias{ggraph}
\title{Create a quick scatter plot in ggplot.}
\usage{
ggraph(x, y, point_color = "black", point_size = 1.5, point_shape = 19)
}
\arguments{
\item{x}{This is the first vector to be plotted.}

\item{y}{This is the first vector to be plotted.}

\item{point_color}{This is the color of the points that will be plotted.}

\item{point_size}{This is the size of the points that will be plotted.
The default is size 1.5.}

\item{point_shape}{This is the shape of the points that will be plotted.
The default is 19: a filled circle.}
}
\value{
This function returns a ggplot scatter plot object.
}
\description{
This will graph two given vectors in a ggplot-style scatter plot with the
x-axis labeled "x" and the y-axis labeled "y".
}
\examples{
## Create a scatter plot of y vs x.
x <- rnorm(100)
y <- x + rnorm(100, 0, 0.3)
ggraph(x, y)

}
