% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghist.R
\name{ghist}
\alias{ghist}
\title{Create a quick histogram in ggplot.}
\usage{
ghist(x)
}
\arguments{
\item{x}{This is the vector to be plotted.}
}
\value{
This function returns a ggplot histogram object.
}
\description{
This will graph a given vectors in a ggplot-style histogram with the
x-axis labeled "x".
}
\examples{
## Create a histogram of x.
x <- rnorm(100)
ghist(x)

}
