% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_plots.R
\name{influence_plots}
\alias{influence_plots}
\title{Influence Plots}
\usage{
influence_plots(model)
}
\arguments{
\item{model}{This is a lm or glm (with binomial family) object}
}
\value{
This function returns plots for the jackknife (externally
studentized) residuals vs index, jackknife residuals vs fitted values,
leverage values vs index, Cook's distance vs index, DfFits vs index, and all
DfBetas vs index.
}
\description{
This will compute many common residual and influence plots to check
adequacy of a given model.
}
\examples{
## Randomly generate correlated variables x and y and then create the 
## influence plots for them
x <- rnorm(100)
y <- x + rnorm(100, 0, 0.3)
influence_plots(lm(y ~ x))

}
